feat(web): extract saved-search-alerts helpers into saved-search-alerts-lib
Move pure saved-search matching and alert builders into a testable lib module
with comprehensive coverage while keeping the public import path unchanged.
